Add access checks

+++ app/activities/controllers/remove.js
@@ -1,12 +1,35 @@
 const removeService = require('../services/remove');
 const { ok, error } = require('../../../core/util/response');
-
+const accessCheck = require('../../projects/services/accessCheck.js');
 /**
  * @api {delete} /api/activities/:id delete
  * @apiName DeleteActivity
  * @apiGroup Activities
  * @apiVersion 1.0.0
  * @apiDescription Delete an activity
+ *
+ * @apiSuccessExample
+{
+    "status": "ok",
+    "message": {
+        "en": "Activity deleted",
+        "fa": "فعالیت حذف شد"
+    },
+    "result": {
+        "id": 17,
+        "start_date": "2021-07-13T19:30:00.000Z",
+        "scheduled_end_date": "2021-07-13T19:30:00.000Z",
+        "project_id": 1,
+        "people": [
+            3
+        ],
+        "zones": [
+            2
+        ],
+        "status": "3",
+        "is_done": false
+    }
+}
  */
 
 const remove = async (request, response) => {
@@ -16,6 +39,11 @@ const remove = async (request, response) => {
         return error(response, 404, {
             en: 'Activity not found.',
             fa: 'فعالیت یافت نشد.',
+        });
+    }
+    if (!(await accessCheck(request.user, activity.project_id))) {
+        return error(response, 403, {
+            en: 'you don\'t have access to this project',
         });
     }
     await removeService.remove_activity(id);
